refactor: move provinces.ts into lib/data directory
