+++ __tests__/pages/provinces/[provinceSlug]/[contactSlug].test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import { contactBuilder } from "~/lib/__mocks__/builders/provinces";
-import provinces from "~/lib/provinces";
+import { contactBuilder } from "~/lib/data/__mocks__/builders/provinces";
+import provinces from "~/lib/data/provinces";
 import ContactPage, {
   getStaticPaths,
   getStaticProps,
@@ -10,7 +10,7 @@ import ContactPage, {
 import { perBuild } from "@jackfranklin/test-data-bot";
 import { render, screen } from "@testing-library/react";
 
-jest.mock("~/lib/provinces");
+jest.mock("~/lib/data/provinces");
 
 describe("ContactPage", () => {
   const [province] = provinces;
