+++ __tests__/pages/provinces/[provinceSlug]/index.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import provinces from "~/lib/provinces";
+import provinces from "~/lib/data/provinces";
 import ProvincePage, {
   getStaticPaths,
   getStaticProps,
@@ -9,7 +9,7 @@ import ProvincePage, {
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-jest.mock("~/lib/provinces");
+jest.mock("~/lib/data/provinces");
 jest.mock("next/router", () => require("next-router-mock"));
 
 describe("ProvincePage", () => {
