+++ pages/provinces/[provinceSlug]/[contactSlug].tsx
@@ -5,9 +5,9 @@ import { Page } from "~/components/layout/page";
 import { PageContent } from "~/components/layout/page-content";
 import { PageHeader } from "~/components/layout/page-header";
 import { ReportButton } from "~/components/report-button";
+import { getContactsPaths } from "~/lib/data/helpers/provinces";
+import provinces, { Contact } from "~/lib/data/provinces";
 import { getContactMeta } from "~/lib/meta";
-import { getContactsPaths } from "~/lib/province-utils";
-import provinces, { Contact } from "~/lib/provinces";
 
 import { GetStaticPaths, GetStaticProps } from "next";
 import { NextSeo } from "next-seo";
