+++ pages/provinces/[provinceSlug]/index.tsx
@@ -6,11 +6,11 @@ import { PageContent } from "~/components/layout/page-content";
 import { PageHeader } from "~/components/layout/page-header";
 import { SearchForm } from "~/components/search-form";
 import { SeoText } from "~/components/seo-text";
+import { getProvincesPaths } from "~/lib/data/helpers/provinces";
+import provinces, { Contact } from "~/lib/data/provinces";
 import { getCurrentLongDate } from "~/lib/date-utils";
 import { useSearch } from "~/lib/hooks/use-search";
 import { getProvinceMeta } from "~/lib/meta";
-import { getProvincesPaths } from "~/lib/province-utils";
-import provinces, { Contact } from "~/lib/provinces";
 
 import { GetStaticPaths, GetStaticProps } from "next";
 import { NextSeo } from "next-seo";
