+++ pages/provinces/index.tsx
@@ -7,9 +7,9 @@ import { LinksWell } from "~/components/links-well";
 import { ProvinceList, ProvinceListItem } from "~/components/province-list";
 import { SearchForm } from "~/components/search-form";
 import { SeoText } from "~/components/seo-text";
+import provinces from "~/lib/data/provinces";
 import { getCurrentMonthAndYear } from "~/lib/date-utils";
 import { useSearch } from "~/lib/hooks/use-search";
-import provinces from "~/lib/provinces";
 import { getInitial } from "~/lib/string-utils";
 
 import { GetStaticProps } from "next";
